+++ packages/http-client-csharp/emitter/src/options.ts
@@ -1,6 +1,7 @@
 import { CreateSdkContextOptions } from "@azure-tools/typespec-client-generator-core";
 import { EmitContext, JSONSchemaType } from "@typespec/compiler";
 import { _defaultGeneratorName } from "./constants.js";
+import { CSharpEmitterContext } from "./index.js";
 import { LoggerLevel } from "./lib/logger-level.js";
 import { CodeModel } from "./type/code-model.js";
 
@@ -18,7 +19,7 @@ export interface CSharpEmitterOptions {
   "disable-xml-docs"?: boolean;
   "generator-name"?: string;
   "emitter-extension-path"?: string;
-  "update-code-model"?: (model: CodeModel) => CodeModel;
+  "update-code-model"?: (model: CodeModel, context: CSharpEmitterContext) => CodeModel;
   "sdk-context-options"?: CreateSdkContextOptions;
   "generate-protocol-methods"?: boolean;
   "generate-convenience-methods"?: boolean;
@@ -158,7 +159,7 @@ export const defaultOptions = {
   debug: undefined,
   logLevel: LoggerLevel.INFO,
   "generator-name": _defaultGeneratorName,
-  "update-code-model": (model: CodeModel) => model,
+  "update-code-model": (model: CodeModel, context: CSharpEmitterContext) => model,
   "sdk-context-options": undefined,
 };