+++ packages/http-client-python/generator/test/unittests/test_parameter_ordering.py
@@ -23,11 +23,11 @@ def get_code_model():
             "namespace": "namespace",
         },
         options={
-            "show_send_request": True,
-            "builders_visibility": "public",
-            "show_operations": True,
-            "models_mode": "dpg",
-            "only_path_and_body_params_positional": True,
+            "show-send-request": True,
+            "builders-visibility": "public",
+            "show-operations": True,
+            "models-mode": "dpg",
+            "only-path-and-body-params-positional": True,
         },
     )